capture request from user to load_balancer and server to  load_balancer is done

+++ Model/model.go
@@ -42,14 +42,14 @@ func (b *Backend) Check(ip string) *Server {
 	return nil
 }
 
-var portCount uint64 = 49152
+var portCount uint16 = 49152
 
 type IpAndPort struct {
 	IP   string
-	Port uint64
+	Port uint16
 }
 
-func UpdatePort(mu *sync.Mutex) uint64 {
+func UpdatePort(mu *sync.Mutex) uint16 {
 	mu.Lock()
 	portCount = portCount + 1
 	if portCount == 65535 {
@@ -60,6 +60,6 @@ func UpdatePort(mu *sync.Mutex) uint64 {
 	return portCount
 }
 
-func GetPort() uint64 {
+func GetPort() uint16 {
 	return portCount
 }
